+++ src/frontend/src/stores/benchmark.js
@@ -52,6 +52,8 @@ export const useBenchmarkStore = defineStore('benchmark', {
 
     // Notification
     notification: { type: null, message: null },
+    // Sys Info modal
+    showSysInfo: false,
   }),
 
   getters: {